integration-structure/project-agro/sistema-agropecuario/backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.4
celery==5.3.4
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
Django==4.2.16
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.3.1
iniconfig==2.3.0
kombu==5.6.1
lxml==6.0.2
msgpack==1.1.2
nfelib==2.2.1
numpy==2.2.6
packaging==25.0
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
PyPDF2==3.0.1
pypdfium2==5.2.0
pypng==0.20220715.0
pytesseract==0.3.10
pytest==9.0.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzbar==0.1.9
qrcode==7.4.2
redis==5.0.1
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.3
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
xsdata==25.7
signxml==4.2.2
